Fix bug coloring an entire row
There was this bug that entire row was getting colored, when mouse was
hovered over a div often due to applying of "colorIt" class on the divs
with "row" class. Earlier, entire grid was getting colored due to same
issue with the "container" class. Fixed this issue, by applying "colorIt"
class only on to "column" class divs i.e. single boxes.